requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
torch==2.0.1
transformers==4.30.2
accelerate==0.20.3
protobuf==3.20.3 